requirements.txt
Flask==2.3.3
spotipy==2.23.0
ytmusicapi==1.10.3
python-dotenv==1.0.0